chore(run): Create generic configuration for RunView
