+++ src/components/data-workflows/common/ViewConversation.vue
@@ -1,0 +1,9 @@
+<template>
+	<blockquote class="blockquote">Coming soon...</blockquote>
+</template>
+
+<script>
+export default {
+	name: 'view-conversation'
+};
+</script>
