+++ src/components/data-workflows/gbq-to-gcs/configuration/overview-data.js
@@ -3,7 +3,7 @@ export default {
 		overviewData() {
 			return [
 				{
-					component: 'overview-header',
+					component: 'view-header',
 					props: {
 						item: this.conf,
 						collection: 'getGbqToGcsConfs',
