+++ src/views/data-workflows/gbq-to-gcs/tab-items.js
@@ -1,6 +1,6 @@
 export default {
 	computed: {
-		tabsItems() {
+		headerTabsItems() {
 			return [
 				{ id: 1, title: 'Runs', link: { name: 'gbqToGcsRuns' } },
 				{ id: 2, title: 'Configurations', link: { name: 'gbqToGcsConfs' } },
