+++ src/views/data-workflows/table-to-storage/configuration/ConfigurationView.vue
@@ -1,6 +1,6 @@
 <template>
 	<div>
-		<data-management-header :tabsItems="tabsItems" />
+		<data-management-header :tabsItems="headerTabsItems" />
 		<configuration-item />
 	</div>
 </template>
@@ -9,7 +9,7 @@
 import DataManagementHeader from '../../../../components/app/headers/DataManagementHeader';
 import ConfigurationItem from '../../../../components/data-workflows/table-to-storage/configuration/ConfigurationItem';
 
-import TabItemsMixin from '../tab-items';
+import TabItemsMixin from '../header-tabs-items';
 
 export default {
 	name: 'table-to-storage-configuration-item',
