+++ src/views/data-workflows/table-to-storage/header-tabs-items.js
@@ -1,6 +1,6 @@
 export default {
 	computed: {
-		tabsItems() {
+		headerTabsItems() {
 			return [
 				{ id: 1, title: 'Runs', link: { name: 'TableToStorageRuns' } },
 				{ id: 2, title: 'Configurations', link: { name: 'TableToStorageConfs' } },
